requirements.txt
alembic==1.10.3
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
dnspython==2.3.0
email-validator==2.0.0.post1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
swagger-ui-bundle==0.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1